requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0

# Data Sources
ib_insync>=0.9.85  # Interactive Brokers API
polygon-api-client>=1.12.3  # Polygon.io API
tradingview-ta>=3.3.0  # TradingView Technical Analysis

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL for production
aiosqlite>=0.19.0  # SQLite for development

# Messaging
discord-webhook>=1.3.0
python-telegram-bot>=20.7
aiosmtplib>=3.0.1  # For email notifications

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
aiohttp>=3.9.1
websockets>=12.0
loguru>=0.7.2 